+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   Patch 11.0.0, updating dependencies
+-   Taint Fix: you are not suppose to move stuff while in combat.
 
 ## [1.0.0] - 2024-07-27